requirements.txt
django==5.2.6
djangorestframework==3.16.1
pytest-django==4.8.0